advertising table + seeder

+++ database/migrations/2022_04_11_151539_create_advertisings_table.php
@@ -15,7 +15,7 @@ class CreateAdvertisingsTable extends Migration
     {
         Schema::create('advertisings', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('duration');
+            $table->smallInteger('duration');
             $table->float('price');
             $table->timestamps();
         });
